Fix progress reporting for dedup.
git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/branches/mapred@320899 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/crawl/DeleteDuplicates.java
@@ -127,7 +127,7 @@ public class DeleteDuplicates extends NutchConfigured
   }
 
   public static class InputFormat extends InputFormatBase {
-    private static final int INDEX_LENGTH = Integer.MAX_VALUE;
+    private static final long INDEX_LENGTH = Integer.MAX_VALUE;
 
     /** Return each index as a split. */
     public FileSplit[] getSplits(NutchFileSystem fs, JobConf job,
